+++ include/mdk/c/VideoFrame.h
@@ -55,7 +55,7 @@ typedef struct mdkVideoFrameAPI {
     double (*timestamp)(struct mdkVideoFrame*);
 
     struct mdkVideoFrameAPI* (*to)(struct mdkVideoFrame*, enum MDK_PixelFormat format, int width/*= -1*/, int height/*= -1*/);
-    bool (*save)(struct mdkVideoFrame*, const char* fileName, const char* format, int quality);
+    bool (*save)(struct mdkVideoFrame*, const char* fileName, const char* format, float quality);
 
 /* The followings are not implemented */
     struct mdkVideoFrameAPI* (*toHost)(struct mdkVideoFrame*);
